+++ hw8/src/GraphM.java
@@ -4,7 +4,7 @@
  *
  */
 public class GraphM implements Graph {
-    private int[][]  matrix;
+    private double[][]  matrix;
     private Object[] nodeValues;
     private int      numEdge;
 
@@ -20,7 +20,7 @@ public class GraphM implements Graph {
 
     // Initialize the graph with n vertices
     public void init(int n) {
-        matrix = new int[n][n];
+        matrix = new double[n][n];
         nodeValues = new Object[n];
         numEdge = 0;
     }
@@ -52,7 +52,7 @@ public class GraphM implements Graph {
 
     // Adds a new edge from node v to node w
     // Returns the new edge
-    public void addEdge(int v, int w, int wgt) {
+    public void addEdge(int v, int w, double wgt) {
         if (wgt == 0) {
             return; // Can't store weight of 0
         }
@@ -64,7 +64,7 @@ public class GraphM implements Graph {
 
 
     // Get the weight value for an edge
-    public int weight(int v, int w) {
+    public double weight(int v, int w) {
         return matrix[v][w];
     }
